+++ src/System.Text.Json/tests/WritableJsonApiTests.cs
@@ -2,144 +2,16 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using System.Runtime.InteropServices.ComTypes;
-using System.Text.Json.Serialization.Tests;
 using Xunit;
-
 
 namespace System.Text.Json
 {
 #pragma warning disable xUnit1000
-    internal static class WritableJsonApiTests
+    internal static partial class WritableJsonApiTests
 #pragma warning enable xUnit1000
     {
-        /// <summary>
-        /// Helper class simulating external library
-        /// </summary>
-        private static class EmployeesDatabase
-        {
-            private static int s_id = 0;
-            public static KeyValuePair<string, JsonNode> GetNextEmployee()
-            {
-                var employee = new JsonObject()
-                {
-                    { "name", "John" } ,
-                    { "surname", "Smith"},
-                    { "age", 45 }
-                };
-
-                return new KeyValuePair<string, JsonNode>("employee" + s_id++, employee);
-            }
-
-            public static IEnumerable<KeyValuePair<string, JsonNode>> GetTenBestEmployees()
-            {
-                for (int i = 0; i < 10; i++)
-                    yield return GetNextEmployee();
-            }
-
-            /// <summary>
-            /// Returns following JsonObject:
-            /// {
-            ///     { "name" : "John" }
-            ///     { "phone numbers" : { "work" :  "123-456-7890", "home": "123-456-7890"  } }
-            ///     { 
-            ///         "reporting employees" : 
-            ///         {
-            ///             "software developers" :
-            ///             {
-            ///                 "full time employees" : /JsonObject of 3 employees fromk database/ 
-            ///                 "intern employees" : /JsonObject of 2 employees fromk database/ 
-            ///             },
-            ///             "HR" : /JsonObject of 10 employees fromk database/ 
-            ///         }
-            /// </summary>
-            /// <returns></returns>
-            public static JsonObject GetManager()
-            {
-                var manager = GetNextEmployee().Value as JsonObject;
-
-                manager.Add
-                (
-                    "phone numbers",
-                    new JsonObject()
-                    {
-                        { "work", "123-456-7890" }, { "home", "123-456-7890" }
-                    }
-                );
-
-                manager.Add
-                (
-                    "reporting employees", new JsonObject()
-                    {
-                        {
-                            "software developers", new JsonObject()
-                            {
-                                {
-                                    "full time employees", new JsonObject()
-                                    {
-                                        EmployeesDatabase.GetNextEmployee(),
-                                        EmployeesDatabase.GetNextEmployee(),
-                                        EmployeesDatabase.GetNextEmployee(),
-                                    }
-                                },
-                                {
-                                    "intern employees", new JsonObject()
-                                    {
-                                        EmployeesDatabase.GetNextEmployee(),
-                                        EmployeesDatabase.GetNextEmployee(),
-                                    }
-                                }
-                            }
-                        },
-                        {
-                            "HR", new JsonObject()
-                            {
-                                {
-                                    "full time employees", new JsonObject(EmployeesDatabase.GetTenBestEmployees())
-                                }
-                            }
-                        }
-                    }
-                );
-
-                return manager;
-            }
-
-            public static bool CheckSSN(string ssnNumber) => true;
-            public static void PerformHeavyOperations(JsonElement employee) { }
-        }
-
-        /// <summary>
-        /// Helper class simulating sending Json files via network
-        /// </summary>
-        private static class Mailbox
-        {
-            public static void SendEmployeeData(JsonElement employeeData) { }
-            public static JsonElement RetrieveMutableEmployeeData() { return EmployeesDatabase.GetNextEmployee().Value.AsJsonElement(); }
-            public static JsonElement RetrieveImmutableEmployeeData() { return new JsonElement(); }
-
-            public static void SendAllEmployeesData(JsonElement employeesData) { }
-        }
-
-        private static class HealthCare
-        {
-            public static void CreateMedicalAppointment(string personName) { }
-        }
-
-        /// <summary>
-        /// Helper class simulating enum
-        /// </summary>
-        private enum AvailableStateCodes
-        {
-            WA,
-            CA,
-            NY,
-        }
-
         /// <summary>
         /// Creating simple Json object
         /// </summary>
@@ -559,28 +431,6 @@ namespace System.Text.Json
             Assert.IsType<JsonArray>(innerJsonArray[0]);
         }
 
-        /// <summary>
-        /// Adding values to JsonArray
-        /// </summary>
-        [Fact]
-        public static void TestAddingToJsonArray()
-        {
-            var employeesIds = new JsonArray();
-
-            foreach (KeyValuePair<string, JsonNode> employee in EmployeesDatabase.GetTenBestEmployees())
-            {
-                employeesIds.Add(employee.Key);
-            }
-
-            JsonString prevId = new JsonString();
-            foreach(JsonNode employeeId in employeesIds)
-            {
-                Assert.IsType<JsonString>(employeeId);
-                var employeeIdString = employeeId as JsonString;
-                Assert.NotEqual(prevId, employeeIdString);
-                prevId = employeeIdString;
-            }
-        }
 
         /// <summary>
         /// Creating Json array from collection
@@ -666,277 +516,6 @@ namespace System.Text.Json
         }
 
         /// <summary>
-        /// Replacing Json object's primnary types
-        /// </summary>
-        [Fact]
-        public static void TestReplacingsonObjectPrimaryTypes()
-        {
-            var person1 = new JsonObject
-            {
-                { "name", "John" },
-                { "age", 45 },
-                { "is_married", true }
-            };
-
-            // Assign by creating a new instance of primary Json type
-            person1["name"] = new JsonString("Bob");
-
-            Assert.IsType<JsonString>(person1["name"]);
-            Assert.Equal(person1["name"] as JsonString, "Bob");
-
-            // Assign by using an implicit operator on primary Json type
-            JsonNumber newAge = 55;
-            person1["age"] = newAge;
-
-            Assert.IsType<JsonNumber>(person1["age"]);
-            Assert.Equal(person1["age"] as JsonNumber, 55);
-
-            // Assign by explicit cast from Json primary type
-            person1["is_married"] = (JsonBoolean)false;
-
-            Assert.IsType<JsonBoolean>(person1["is_married"]);
-            Assert.Equal(person1["is_married"] as JsonBoolean, false);
-
-            // Not possible scenario (wold require implicit cast operators in JsonNode):
-            // person["name"] = "Bob";
-
-            var person2 = new JsonObject
-            {
-                { "name", new JsonString[]{ "Emily", "Rosalie" } },
-                { "age", 33 },
-                { "is_married", true }
-            };
-
-            // Copy property from another JsonObject
-            person1["age"] = person2["age"];
-
-            Assert.IsType<JsonNumber>(person1["age"]);
-            Assert.Equal(person1["age"] as JsonNumber, 33);
-
-            // Copy property of different typoe
-            person1["name"] = person2["name"];
-
-            Assert.IsType<JsonArray>(person1["name"]);
-        }
-
-        /// <summary>
-        /// Modifying Json object's primnary types
-        /// </summary>
-        [Fact]
-        public static void TestModifyingJsonObjectPrimaryTypes()
-        {
-            JsonString name = "previous name";
-            name.Value = "new name";
-
-            Assert.Equal("new name", name.Value);
-
-            bool shouldBeEnabled = true;
-            var isEnabled = new JsonBoolean(false);
-            isEnabled.Value = shouldBeEnabled;
-
-            Assert.True(isEnabled.Value);
-
-            JsonNumber veryBigConstant = new JsonNumber();
-            veryBigConstant.SetString("1e1000");
-            string bigNumber = veryBigConstant.GetString();
-
-            Assert.Equal("1e1000", bigNumber);
-
-            veryBigConstant.SetInt16(123);
-            short smallNumber = veryBigConstant.GetInt16();
-
-            Assert.Equal(123, smallNumber);
-
-            // Incrementing age:
-            JsonObject employee = EmployeesDatabase.GetManager();
-            int age = ((JsonNumber)employee["age"]).GetInt32();
-            ((JsonNumber)employee["age"]).SetInt32(age + 1);
-
-            Assert.Equal(46, ((JsonNumber)employee["age"]).GetInt32());
-        }
-
-        /// <summary>
-        /// Accesing nested Json object - casting with as operator
-        /// </summary>
-        [Fact]
-        public static void TestAccesingNestedJsonObjectCastWithAs()
-        {
-            JsonObject manager = EmployeesDatabase.GetManager();
-
-            // Should not throw any exceptions:
-
-            var reportingEmployees = manager["reporting employees"] as JsonObject;
-            if (reportingEmployees == null)
-                throw new InvalidCastException();
-
-            var softwareDevelopers = reportingEmployees["software developers"] as JsonObject;
-            if (softwareDevelopers == null)
-                throw new InvalidCastException();
-
-            var internDevelopers = softwareDevelopers["intern employees"] as JsonObject;
-            if (internDevelopers == null)
-                throw new InvalidCastException();    
-
-            internDevelopers.Add(EmployeesDatabase.GetNextEmployee());
-        }
-
-        /// <summary>
-        /// Accesing nested Json object - casting with is operator
-        /// </summary>
-        [Fact]
-        public static void TestAccesingNestedJsonObjectCastWithIs()
-        {
-            JsonObject manager = EmployeesDatabase.GetManager();
-
-            static bool AddEmployee(JsonObject manager)
-            {
-
-                if (manager["reporting employees"] is JsonObject reportingEmployees)
-                {
-                    if (reportingEmployees["software developers"] is JsonObject softwareDevelopers)
-                    {
-                        if (softwareDevelopers["full time employees"] is JsonObject fullTimeEmployees)
-                        {
-                            fullTimeEmployees.Add(EmployeesDatabase.GetNextEmployee());
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            bool success = AddEmployee(manager);
-            Assert.True(success);
-        }
-
-        /// <summary>
-        /// Accesing nested Json object - explicit casting
-        /// </summary>
-        [Fact]
-        public static void TestAccesingNestedJsonObjectExplicitCast()
-        {
-            JsonObject manager = EmployeesDatabase.GetManager();
-
-            // Should not throw any exceptions:
-            ((JsonObject)((JsonObject)manager["reporting employees"])["HR"]).Add(EmployeesDatabase.GetNextEmployee());
-        }
-
-        /// <summary>
-        /// Accesing nested Json object - GetObjectProperty method
-        /// </summary>
-        [Fact]
-        public static void TestAccesingNestedJsonObjectGetPropertyMethod()
-        {
-            JsonObject manager = EmployeesDatabase.GetManager();
-
-            // Should not throw any exceptions:
-
-            JsonObject internDevelopers = manager.GetObjectProperty("reporting employees")
-                                          .GetObjectProperty("software developers")
-                                          .GetObjectProperty("intern employees");
-            internDevelopers.Add(EmployeesDatabase.GetNextEmployee());
-        }
-
-        /// <summary>
-        /// Accesing nested Json object - TryGetObjectProperty method
-        /// </summary>
-        [Fact]
-        public static void TestAccesingNestedJsonObjectTryGetPropertyMethod()
-        {
-            JsonObject manager = EmployeesDatabase.GetManager();
-
-            static bool AddEmployee(JsonObject manager)
-            {
-                if (manager.TryGetObjectProperty("reporting employees", out JsonObject reportingEmployees))
-                {
-                    if (reportingEmployees.TryGetObjectProperty("software developers", out JsonObject softwareDevelopers))
-                    {
-                        if (softwareDevelopers.TryGetObjectProperty("full time employees", out JsonObject fullTimeEmployees))
-                        {
-                            fullTimeEmployees.Add(EmployeesDatabase.GetNextEmployee());
-                            return true;
-                        }
-                    }
-                }
-
-                return false;
-            }
-
-            bool success = AddEmployee(manager);
-            Assert.True(success);
-        }
-
-        /// <summary>
-        /// Accesing nested Json array - GetArrayProperty method
-        /// </summary>
-        [Fact]
-        public static void TestAccesingNestedJsonArrayGetPropertyMethod()
-        {
-            var issues = new JsonObject()
-            {
-                { "features", new JsonArray{ "new functionality 1", "new functionality 2" } },
-                { "bugs", new JsonArray{ "bug 123", "bug 4566", "bug 821" } },
-                { "tests", new JsonArray{ "code coverage" } },
-            };
-
-            issues.GetArrayProperty("bugs").Add("bug 12356");
-            ((JsonString)issues.GetArrayProperty("features")[0]).Value = "feature 1569";
-            ((JsonString)issues.GetArrayProperty("features")[1]).Value = "feature 56134";
-
-            Assert.True(((JsonArray)issues["bugs"]).Contains("bug 12356"));
-            Assert.Equal((JsonString)((JsonArray)issues["features"])[0], "feature 1569");
-            Assert.Equal((JsonString)((JsonArray)issues["features"])[1], "feature 56134");
-        }
-
-        /// <summary>
-        /// Modifying Json object key - remove and add
-        /// </summary>
-        [Fact]
-        public static void TestModifyingJsonObjectKeyRemoveAdd()
-        {
-            JsonObject manager = EmployeesDatabase.GetManager();
-            JsonObject reportingEmployees = manager.GetObjectProperty("reporting employees");
-
-            static void ModifyProperty(JsonObject jsonObject, string previousName, string newName)
-            {
-                JsonNode previousValue = jsonObject[previousName];
-                jsonObject.Remove(previousName);
-                jsonObject.Add(newName, previousValue);
-            }
-
-            string previousName = "software developers";
-            string newName = "software engineers";
-
-            Assert.True(reportingEmployees.ContainsProperty(previousName));
-            JsonNode previousValue = reportingEmployees[previousName];
-
-            ModifyProperty(reportingEmployees, previousName, newName);
-
-            Assert.False(reportingEmployees.ContainsProperty(previousName));
-            Assert.True(reportingEmployees.ContainsProperty(newName));
-            Assert.Equal(previousValue, reportingEmployees[newName]);
-        }
-
-        /// <summary>
-        /// Modifying Json object key - modify method
-        /// </summary>
-        [Fact]
-        public static void TestModifyingJsonObjectKeyModifyMethod()
-        {
-            JsonObject manager = EmployeesDatabase.GetManager();
-            JsonObject reportingEmployees = manager.GetObjectProperty("reporting employees");
-
-            Assert.True(reportingEmployees.ContainsProperty("software developers"));
-            JsonNode previousValue = reportingEmployees["software engineers"];
-
-            reportingEmployees.ModifyPropertyName("software developers", "software engineers");
-
-            Assert.False(reportingEmployees.ContainsProperty("software developers"));
-            Assert.True(reportingEmployees.ContainsProperty("software engineers"));
-            Assert.Equal(previousValue, reportingEmployees["software engineers"]);
-        }
-
-        /// <summary>
         /// Aquiring all values
         /// </summary>
         [Fact]
@@ -971,223 +550,6 @@ namespace System.Text.Json
             Assert.True(fullTimeEmployees.Contains((JsonString)"John Smith"));
             Assert.True(fullTimeEmployees.Contains((JsonString)"Ann Predictable"));
             Assert.True(fullTimeEmployees.Contains((JsonString)"Byron Shadow"));
-        }
-
-        /// <summary>
-        /// Transforming JsoneNode to JsonElement
-        /// </summary>
-        [Fact]
-        public static void TestTransformingJsonNodeToJsonElement()
-        {
-            // Send Json through network, should not throw any exceptions:
-            JsonNode employeeDataToSend = EmployeesDatabase.GetNextEmployee().Value;
-            Mailbox.SendEmployeeData(employeeDataToSend.AsJsonElement());
-        }
-
-        /// <summary>
-        /// Transforming JsonElement to JsonNode
-        /// </summary>
-        [Fact]
-        public static void TestTransformingJsonElementToJsonNode()
-        {
-            // Retrieve Json from network, should not throw any exceptions:
-            JsonNode receivedEmployeeData = JsonNode.DeepCopy(Mailbox.RetrieveMutableEmployeeData());
-            if (receivedEmployeeData is JsonObject employee)
-            {
-                employee["name"] = new JsonString("Bob");
-                Mailbox.SendEmployeeData(employee.AsJsonElement());
-            }
-        }
-
-        /// <summary>
-        /// Transforming JsonDocument to JsonNode and vice versa
-        /// </summary>
-        [Fact]
-        public static void TestTransformingToFromJsonDocument()
-        {
-            string jsonString = @"
-            {
-                ""employee1"" : 
-                {
-                    ""name"" : ""Ann"",
-                    ""surname"" : ""Predictable"",
-                    ""age"" : 30,                
-                },
-                ""employee2"" : 
-                {
-                    ""name"" : ""Zoe"",
-                    ""surname"" : ""Coder"",
-                    ""age"" : 24,                
-                }
-            }";
-
-            using (JsonDocument employeesToSend = JsonDocument.Parse(jsonString))
-            {
-                // regular send:
-                Mailbox.SendAllEmployeesData(employeesToSend.RootElement);
-
-                // modified elements send:
-                JsonObject employees = JsonNode.DeepCopy(employeesToSend) as JsonObject;
-                Assert.Equal(2, employees.Count());
-
-                employees.Add(EmployeesDatabase.GetNextEmployee());
-                Assert.Equal(3, employees.Count());
-
-                Mailbox.SendAllEmployeesData(employees.AsJsonElement());
-            }
-        }
-
-        /// <summary>
-        /// Parsing right to JsonNode if user knows data will be modified
-        /// </summary>
-        [Fact]
-        public static void TestParsingToJsonNode()
-        {
-            string jsonString = @"
-            {
-                ""employee1"" : 
-                {
-                    ""name"" : ""Ann"",
-                    ""surname"" : ""Predictable"",
-                    ""age"" : 30,                
-                },
-                ""employee2"" : 
-                {
-                    ""name"" : ""Zoe"",
-                    ""surname"" : ""Coder"",
-                    ""age"" : 24,                
-                }
-            }";
-
-            JsonObject employees = JsonNode.Parse(jsonString) as JsonObject;
-            employees.Add(EmployeesDatabase.GetNextEmployee());
-            Mailbox.SendAllEmployeesData(employees.AsJsonElement());
-
-            Assert.Equal(2, employees.Count());
-            Assert.True(employees.ContainsProperty("employee1"));
-            Assert.True(employees.ContainsProperty("employee2"));
-
-            JsonObject employee2 = employees.GetObjectProperty("employee2");
-            Assert.IsType<JsonString>(employee2["name"]);
-            Assert.Equal("Zoe", (JsonString)employee2["name"]);
-            Assert.IsType<JsonString>(employee2["surname"]);
-            Assert.Equal("Coder", (JsonString)employee2["surname"]);
-            Assert.IsType<JsonNumber>(employee2["age"]);
-            Assert.Equal(24, (JsonNumber)employee2["age"]);
-        }
-
-        /// <summary>
-        /// Copying JsoneNode
-        /// </summary>
-        [Fact]
-        public static void TestCopyingJsonNode()
-        {
-            JsonObject employee = EmployeesDatabase.GetManager();
-            JsonNode employeeCopy = JsonNode.DeepCopy(employee);
-
-            static bool RecursiveEquals(JsonNode left, JsonNode right)
-            {
-                if (left == null && right == null)
-                {
-                    return true;
-                }
-               
-                if(right.GetType() != left.GetType())
-                {
-                    return false;
-                }
-
-                switch(left)
-                {
-                    case JsonObject leftJsonObject:
-                        var rightJsonObject = right as JsonObject;
-                        if (leftJsonObject.Count() != rightJsonObject.Count())
-                        {
-                            return false;
-                        }
-
-                        for(int idx = 0; idx < leftJsonObject.Count(); idx++)
-                        {
-                            KeyValuePair<string, JsonNode> leftElement = leftJsonObject.ElementAt(idx);
-                            KeyValuePair<string, JsonNode> rightElement = rightJsonObject.ElementAt(idx);
-
-                            if(leftElement.Key != rightElement.Key || !RecursiveEquals(leftElement.Value, rightElement.Value))
-                            {
-                                return false;
-                            }
-                        }
-
-                        return true;
-                    case JsonArray leftJsonArray:
-                        var rightJsonArray = right as JsonArray;
-                        if (leftJsonArray.Count() != rightJsonArray.Count())
-                        {
-                            return false;
-                        }
-                        for (int idx = 0; idx < leftJsonArray.Count(); idx++)
-                        {
-                            JsonNode leftElement = leftJsonArray.ElementAt(idx);
-                            JsonNode rightElement = rightJsonArray.ElementAt(idx);
-
-                            if(!RecursiveEquals(leftElement, rightElement))
-                            {
-                                return false;
-                            }
-                        }
-
-                        return true;
-                    case JsonString leftJsonString:
-                        return leftJsonString.Equals(right as JsonString);
-                    case JsonNumber leftJsonNumber:
-                        return leftJsonNumber.Equals(right as JsonNumber);
-                    case JsonBoolean leftJsonBoolean:
-                        return leftJsonBoolean.Equals(right as JsonBoolean);
-                    default:
-                        return false;
-                }
-            }
-
-            Assert.True(RecursiveEquals(employee, employeeCopy));
-        }
-
-        /// <summary>
-        /// Checking IsImmutable property
-        /// </summary>
-        [Fact]
-        public static void TestIsImmutable()
-        {
-            JsonElement employee = Mailbox.RetrieveMutableEmployeeData();
-            Assert.False(employee.IsImmutable);
-
-            if (!employee.IsImmutable)
-            {
-                JsonObject employeeNode = JsonNode.GetNode(employee) as JsonObject;
-                employeeNode["received as node"] = (JsonBoolean) true;
-            }
-
-            employee = Mailbox.RetrieveImmutableEmployeeData();
-            Assert.True(employee.IsImmutable);
-        }
-
-        /// <summary>
-        /// Retrieving JsonNode from JsonElement
-        /// </summary>
-        [Fact]
-        public static void TestTryGetNode()
-        {
-            JsonElement employee = Mailbox.RetrieveMutableEmployeeData();
-
-            static bool CheckIfReceivedAsNode(JsonElement employee)
-            {
-                if (JsonNode.TryGetNode(employee, out JsonNode employeeNode))
-                {
-                    ((JsonObject)employeeNode)["received as node"] = (JsonBoolean)true;
-                    return true;
-                }
-                return false;
-            }
-
-            Assert.True(CheckIfReceivedAsNode(employee));
         }
 
         /// <summary>
